+++ src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/custom.py
@@ -31,7 +31,8 @@ from azure.mgmt.web.models import (Site, SiteConfig, User, AppServicePlan, SiteC
                                    BackupRequest, DatabaseBackupSetting, BackupSchedule,
                                    RestoreRequest, FrequencyUnit, Certificate, HostNameSslState,
                                    RampUpRule, UnauthenticatedClientAction, ManagedServiceIdentity,
-                                   DeletedAppRestoreRequest, DefaultErrorResponseException)
+                                   DeletedAppRestoreRequest, DefaultErrorResponseException,
+                                   SnapshotRestoreRequest, SnapshotRecoverySource)
 
 from azure.cli.core.commands.client_factory import get_mgmt_service_client
 from azure.cli.core.commands import LongRunningOperation
@@ -39,7 +40,7 @@ from azure.cli.core.util import in_cloud_console
 from azure.cli.core.util import open_page_in_browser
 
 from .vsts_cd_provider import VstsContinuousDeliveryProvider
-from ._params import AUTH_TYPES, MULTI_CONTAINER_TYPES
+from ._params import AUTH_TYPES, MULTI_CONTAINER_TYPES, LINUX_RUNTIMES, WINDOWS_RUNTIMES
 from ._client_factory import web_client_factory, ex_handler_factory
 from ._appservice_utils import _generic_site_operation
 from ._create_util import (zip_contents_from_dir, get_runtime_version_details, create_resource_group,
@@ -1219,6 +1220,39 @@ def restore_backup(cmd, resource_group_name, webapp_name, storage_account_url, b
     return client.web_apps.restore(resource_group_name, webapp_name, 0, restore_request)
 
 
+def list_snapshots(cmd, resource_group_name, name, slot=None):
+    return _generic_site_operation(cmd.cli_ctx, resource_group_name, name, 'list_snapshots',
+                                   slot)
+
+
+def restore_snapshot(cmd, resource_group_name, name, time, slot=None, restore_content_only=False,
+                     source_resource_group=None, source_name=None, source_slot=None):
+    from azure.cli.core.commands.client_factory import get_subscription_id
+    client = web_client_factory(cmd.cli_ctx)
+    recover_config = not restore_content_only
+    if all([source_resource_group, source_name]):
+        # Restore from source app to target app
+        sub_id = get_subscription_id(cmd.cli_ctx)
+        source_id = "/subscriptions/" + sub_id + "/resourceGroups/" + source_resource_group + \
+            "/providers/Microsoft.Web/sites/" + source_name
+        if source_slot:
+            source_id = source_id + "/slots/" + source_slot
+        source = SnapshotRecoverySource(id=source_id)
+        request = SnapshotRestoreRequest(overwrite=False, snapshot_time=time, recovery_source=source,
+                                         recover_configuration=recover_config)
+        if slot:
+            return client.web_apps.restore_snapshot_slot(resource_group_name, name, request, slot)
+        return client.web_apps.restore_snapshot(resource_group_name, name, request)
+    elif any([source_resource_group, source_name]):
+        raise CLIError('usage error: --source-resource-group and --source-name must both be specified if one is used')
+    else:
+        # Overwrite app with its own snapshot
+        request = SnapshotRestoreRequest(overwrite=True, snapshot_time=time, recover_configuration=recover_config)
+        if slot:
+            return client.web_apps.restore_snapshot_slot(resource_group_name, name, request, slot)
+        return client.web_apps.restore_snapshot(resource_group_name, name, request)
+
+
 # pylint: disable=inconsistent-return-statements
 def _create_db_setting(db_name, db_type, db_connection_string):
     if all([db_name, db_type, db_connection_string]):
@@ -1806,7 +1840,7 @@ def create_function(cmd, resource_group_name, name, storage_account, plan=None,
                     os_type=None, runtime=None, consumption_plan_location=None,
                     deployment_source_url=None, deployment_source_branch='master',
                     deployment_local_git=None, deployment_container_image_name=None, tags=None):
-    # pylint: disable=too-many-statements
+    # pylint: disable=too-many-statements, too-many-branches
     if deployment_source_url and deployment_local_git:
         raise CLIError('usage error: --deployment-source-url <url> | --deployment-local-git')
     if bool(plan) == bool(consumption_plan_location):
@@ -1844,13 +1878,21 @@ def create_function(cmd, resource_group_name, name, storage_account, plan=None,
         functionapp_def.server_farm_id = plan
         functionapp_def.location = location
 
+    if runtime:
+        if is_linux and runtime not in LINUX_RUNTIMES:
+            raise CLIError("usage error: Currently supported runtimes (--runtime) in linux function apps are: {}."
+                           .format(', '.join(LINUX_RUNTIMES)))
+        elif not is_linux and runtime not in WINDOWS_RUNTIMES:
+            raise CLIError("usage error: Currently supported runtimes (--runtime) in windows function apps are: {}."
+                           .format(', '.join(WINDOWS_RUNTIMES)))
+        site_config.app_settings.append(NameValuePair(name='FUNCTIONS_WORKER_RUNTIME', value=runtime))
+
     con_string = _validate_and_get_connection_string(cmd.cli_ctx, resource_group_name, storage_account)
 
     if is_linux:
         functionapp_def.kind = 'functionapp,linux'
         functionapp_def.reserved = True
         if consumption_plan_location:
-            site_config.app_settings.append(NameValuePair(name='FUNCTIONS_WORKER_RUNTIME', value=runtime))
             site_config.app_settings.append(NameValuePair(name='FUNCTIONS_EXTENSION_VERSION', value='~2'))
         else:
             site_config.app_settings.append(NameValuePair(name='FUNCTIONS_EXTENSION_VERSION', value='beta'))