+++ src/command_modules/azure-cli-acs/HISTORY.rst
@@ -3,9 +3,16 @@
 Release History
 ===============
 
+2.3.14
+++++++
+* Add Virtual Nodes Preview
+
 2.3.13
 ++++++
 * Remove "(PREVIEW)" from AAD arguments to "az aks create"
+* Mark "az acs" commands as deprecated (the ACS service will retire on January 31, 2020)
+* Add support of Network Policy when creating new AKS clusters
+* Don't require --nodepool-name in "az aks scale" if there's only one nodepool
 
 2.3.12
 ++++++
@@ -33,7 +40,7 @@ Release History
 
 2.3.6
 +++++
-* `az aks create/scale --nodepool-name` configures nodepool name, truncated to 12 characters, default - nodepool1 
+* `az aks create/scale --nodepool-name` configures nodepool name, truncated to 12 characters, default - nodepool1
 
 2.3.5
 +++++