+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -2,9 +2,15 @@
 
 Release History
 ===============
+0.2.10
+++++++
+* webapp: add support for listing and restoring webapp snapshots
+
 0.2.9
 +++++
 * webapp: az webapp config container now honors --slot parameter
+* functionapp: add support for --runtime flag in windows function apps
+
 
 0.2.8
 +++++